+++ src/components/extras/organisms/reference/business-affiliates/supplier-form/supplier-form.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import type { BusinessAffiliate } from '@/types/extras/reference/business-affiliates/business-affiliate';
+import type { BusinessAffiliate } from '@/types/extras';
 import { ArdaSupplierViewer } from '@/components/extras/organisms/reference/business-affiliates/arda-supplier-viewer';
 import { supplierTabs } from '@/components/extras/organisms/reference/business-affiliates/arda-supplier-viewer/configs/stepped-layout';
 import { supplierFieldOrder } from '@/components/extras/organisms/reference/business-affiliates/arda-supplier-viewer/configs/continuous-scroll-layout';