+++ src/components/extras/organisms/reference/items/item-supply-form-dialog/item-supply-form-dialog.stories.tsx
@@ -3,7 +3,7 @@ import { expect, fn, within } from 'storybook/test';
 import { useState } from 'react';
 
 import type { TypeaheadOption } from '@/components/extras/atoms/typeahead/typeahead';
-import { sampleItemSupplies } from '@/types/extras/reference/business-affiliates/item-supply';
+import { sampleItemSupplies } from '../mocks/sample-item-supplies';
 
 import { ArdaItemSupplyFormDialog } from './item-supply-form-dialog';